Add docker-compose override file to use a shared gmail account in datafeeder
The `docker-compose.datafeeder.gmail.yml` file sets up the required
environment variables for datafeeder to use the account
noreply.georchestra.dev@gmail.com to send emails.

The account password is not pushed to github, and must be set in the
`.env` file as `SMTP_PASSWORD=<actual pwd>`.

The `.env` file didn't exist, and has been added to `.gitignore` to
prevent accidental commits sharing the secret.

The `run-datafeeder-gmail.sh` file makes it easier to run the
composition with these settings, checks the env variable is set and runs
the required docker-compose command.